+++ server/binary/marshal.go
@@ -201,7 +201,7 @@ func unmarshalShort(src []byte) (int, int, error) {
 }
 
 // MarshalUInt marshals unsigned 32bit integer
-func MarshalUInt(val int) []byte {
+func MarshalUInt(val int64) []byte {
 	v := clamp(int64(val), 0, math.MaxUint32)
 	buf := make([]byte, 1+UIntDataSize)
 	buf[0] = byte(TypeUInt)
@@ -209,7 +209,7 @@ func MarshalUInt(val int) []byte {
 	return buf
 }
 
-func unmarshalUInt(src []byte) (int, int, error) {
+func unmarshalUInt(src []byte) (int64, int, error) {
 	if len(src) < 1+UIntDataSize {
 		return 0, 0, xerrors.Errorf("Unmarshal UInt error: not enough data (%v)", len(src))
 	}
@@ -217,7 +217,7 @@ func unmarshalUInt(src []byte) (int, int, error) {
 }
 
 // MarshalInt marshals signed 32bit integer comparably
-func MarshalInt(val int) []byte {
+func MarshalInt(val int64) []byte {
 	v := clamp(int64(val), math.MinInt32, math.MaxInt32)
 	buf := make([]byte, 1+IntDataSize)
 	buf[0] = byte(TypeInt)
@@ -225,7 +225,7 @@ func MarshalInt(val int) []byte {
 	return buf
 }
 
-func unmarshalInt(src []byte) (int, int, error) {
+func unmarshalInt(src []byte) (int64, int, error) {
 	if len(src) < 1+IntDataSize {
 		return 0, 0, xerrors.Errorf("Unmarshal Int error: not enough data (%v)", len(src))
 	}
@@ -773,7 +773,7 @@ func unmarshalUShorts(src []byte) ([]int, int, error) {
 //   - TypeInts
 //   - 16bit count
 //   - repeat: 32bit BE integer...
-func MarshalInts(vals []int) []byte {
+func MarshalInts(vals []int64) []byte {
 	if vals == nil {
 		return MarshalNull()
 	}
@@ -787,14 +787,14 @@ func MarshalInts(vals []int) []byte {
 	put16(buf[1:], int64(count))
 
 	for i := 0; i < count; i++ {
-		v := clamp(int64(vals[i]), math.MinInt32, math.MaxInt32) - math.MinInt32
+		v := clamp(vals[i], math.MinInt32, math.MaxInt32) - math.MinInt32
 		put32(buf[3+i*IntDataSize:], v)
 	}
 
 	return buf
 }
 
-func unmarshalInts(src []byte) ([]int, int, error) {
+func unmarshalInts(src []byte) ([]int64, int, error) {
 	if len(src) < 3 {
 		return nil, 0, xerrors.Errorf("Unmarshal Intts error: not enough data (%v)", len(src))
 	}
@@ -803,7 +803,7 @@ func unmarshalInts(src []byte) ([]int, int, error) {
 	if len(src) < l {
 		return nil, 0, xerrors.Errorf("Unmarshal Ints error: not enough data (%v)", len(src))
 	}
-	vals := make([]int, count)
+	vals := make([]int64, count)
 	for i := 0; i < count; i++ {
 		vals[i] = get32(src[3+i*IntDataSize:]) + math.MinInt32
 	}
@@ -815,7 +815,7 @@ func unmarshalInts(src []byte) ([]int, int, error) {
 //   - TypeUInts
 //   - 16bit count
 //   - repeat: 32bit BE integer...
-func MarshalUInts(vals []int) []byte {
+func MarshalUInts(vals []int64) []byte {
 	if vals == nil {
 		return MarshalNull()
 	}
@@ -829,14 +829,14 @@ func MarshalUInts(vals []int) []byte {
 	put16(buf[1:], int64(count))
 
 	for i := 0; i < count; i++ {
-		v := clamp(int64(vals[i]), 0, math.MaxUint32)
+		v := clamp(vals[i], 0, math.MaxUint32)
 		put32(buf[3+i*UIntDataSize:], v)
 	}
 
 	return buf
 }
 
-func unmarshalUInts(src []byte) ([]int, int, error) {
+func unmarshalUInts(src []byte) ([]int64, int, error) {
 	if len(src) < 3 {
 		return nil, 0, xerrors.Errorf("Unmarshal UInts error: not enough data (%v)", len(src))
 	}
@@ -845,7 +845,7 @@ func unmarshalUInts(src []byte) ([]int, int, error) {
 	if len(src) < l {
 		return nil, 0, xerrors.Errorf("Unmarshal UInts error: not enough data (%v)", len(src))
 	}
-	vals := make([]int, count)
+	vals := make([]int64, count)
 	for i := 0; i < count; i++ {
 		vals[i] = get32(src[3+i*UIntDataSize:])
 	}
@@ -1211,11 +1211,11 @@ func put32(dst []byte, val int64) {
 	dst[2] = byte((val & 0xff00) >> 8)
 	dst[3] = byte(val & 0xff)
 }
-func get32(src []byte) int {
-	i := int(src[0]) << 24
-	i += int(src[1]) << 16
-	i += int(src[2]) << 8
-	i += int(src[3])
+func get32(src []byte) int64 {
+	i := int64(src[0]) << 24
+	i += int64(src[1]) << 16
+	i += int64(src[2]) << 8
+	i += int64(src[3])
 	return i
 }
